requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
prophet>=1.1
tensorflow>=2.15
statsmodels>=0.14
sqlalchemy>=2.0
fastapi>=0.110
uvicorn>=0.29
pydantic>=2.7
python-dateutil>=2.9
